+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscanning/config/SwaggerPublisher.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.config;
+package uk.gov.hmcts.reform.bulkscanning.config;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
