+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscanning/controller/GetWelcomeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.controller;
+package uk.gov.hmcts.reform.bulkscanning.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
