+++ src/main/java/uk/gov/hmcts/reform/bulkscanning/BulkScanningApiApplication.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo;
+package uk.gov.hmcts.reform.bulkscanning;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
